eject from CRA, tweak .eslintrc
Ejected from CRA so we could tweak scripts/start.js to use polling.
(note that the edit in start.js is a dirty hack that probbbbably
ought to be in some config file somewhere, but that wasn't working)

This led to trouble with an eslint/babel/class-properties interaction,
and the solution to that was to tell eslint to use babel as its parser.

Also we took the "proxy" setting out of the package.json